Pre-compile 1/255 in more places, don't free nulls, and turn float into double to avoid additional conversion instruction.

+++ HW1/src/triangle.cpp
@@ -49,12 +49,12 @@ double Triangle::getIntersection(Ray ray){
    const auto overX = 1/ray.vector.x;
    const auto overY = 1/ray.vector.y;
    const auto overZ = 1/ray.vector.z;
-   float tmin = (min_v.x - ray.point.x) * overX;
-   float tmax = (max_v.x - ray.point.x) * overX;
+   double tmin = (min_v.x - ray.point.x) * overX;
+   double tmax = (max_v.x - ray.point.x) * overX;
    if (tmin > tmax) std::swap(tmin, tmax);
 
-   float tymin = (min_v.y - ray.point.y) * overY;
-   float tymax = (max_v.y - ray.point.y) * overY;
+   double tymin = (min_v.y - ray.point.y) * overY;
+   double tymax = (max_v.y - ray.point.y) * overY;
    if (tymin > tymax) std::swap(tymin, tymax);
 
    if ((tmin > tymax) || (tymin > tmax))
@@ -65,8 +65,8 @@ double Triangle::getIntersection(Ray ray){
    if (tymax < tmax)
       tmax = tymax;
 
-   float tzmin = (min_v.z - ray.point.z) * overZ;
-   float tzmax = (max_v.z - ray.point.z) * overZ;
+   double tzmin = (min_v.z - ray.point.z) * overZ;
+   double tzmax = (max_v.z - ray.point.z) * overZ;
    if (tzmin > tzmax) std::swap(tzmin, tzmax);
 
    if ((tmin > tzmax) || (tzmin > tmax))
